create a hook and store reservation data
